[Improvement] Refactor PlayerManager
